backend/requirements.txt
alembic==1.17.2
email-validator==2.3.0
fastapi==0.122.0
httpx==0.28.1
pydantic-settings==2.12.0
PyJWT==2.10.1
PyMySQL==1.1.2
pytest==9.0.1
pytest-asyncio==1.3.0
python-dotenv==1.2.1
python-multipart==0.0.20
pwdlib[argon2]==0.2.1
SQLAlchemy==2.0.44
uvicorn[standard]==0.38.0
